Added support for generating partitioned TPC dataset, also added parallelization

+++ datalogistik/tpc_builders.py
@@ -13,12 +13,16 @@
 # limitations under the License.
 
 import abc
+import concurrent.futures
+import multiprocessing
 import os
 import pathlib
 import platform
 import shutil
 import subprocess
 from typing import List, Optional
+
+import tqdm
 
 from .log import log
 from .tpc_info import tpc_table_names
@@ -99,8 +103,15 @@ class _TPCBuilder(abc.ABC):
     def _get_default_executable_path(self):
         """Return the executable path for this generator if one isn't given."""
 
+    @abc.abstractmethod
+    def _get_partitioned_filename(self, table_name: str, p: int, partitions: int):
+        """Create the pathname for a file in case of a partitioned output"""
+
     def create_dataset(
-        self, out_dir: pathlib.Path, scale_factor: int = 1, partitions: int = 1
+        self,
+        out_dir: pathlib.Path,
+        scale_factor: int = 1,
+        partitions: int = 1,
     ):
         """Call the executable to generate the TPC database.
 
@@ -110,35 +121,80 @@ class _TPCBuilder(abc.ABC):
             The directory to place the CSVs in.
         scale_factor
             The scale factor to use when building the database. Default 1.
+        partitions
+            The number of partitions to create, these will be generated in parallel.
+            Default 1.
         """
         if not self.executable_path or not self.executable_path.exists():
             log.info("Could not find an executable. Attempting to create one.")
             self._make_executable()
 
         partitions = 1 if partitions == 0 else partitions
-
-        for thread in range(partitions):
+        if partitions == 1:
+            # TODO: parallelize across files instead of partitions
             _run(
                 self.executable_path,
                 self.force_flag,
                 self.scale_flag,
                 str(scale_factor),
-                self.partitions_flag,
-                str(partitions),
-                self.current_segment_flag,
-                thread,
                 cwd=self.executable_path.parent,
             )
+            # Move the new files to out_dir
+            out_dir = pathlib.Path(out_dir).resolve()
+            for table_name in self.table_names:
+                old_file = self.executable_path.parent / (
+                    table_name + self.file_extension
+                )
+                new_file = out_dir / (table_name + ".csv")
+                shutil.move(old_file, new_file)
+                # reset permissions to read-only
+                os.chmod(new_file, 0o444)
+                log.debug(f"Created {new_file}")
+        else:
+            num_cpus = multiprocessing.cpu_count()
+            pool_size = num_cpus - 1
 
-        # Move the new files to out_dir
-        out_dir = pathlib.Path(out_dir).resolve()
-        for table_name in self.table_names:
-            old_file = self.executable_path.parent / (table_name + self.file_extension)
-            new_file = out_dir / (table_name + ".csv")
-            shutil.move(old_file, new_file)
-            # reset permissions to read-only
-            os.chmod(new_file, 0o444)
-            log.debug(f"Created {new_file}")
+            with concurrent.futures.ProcessPoolExecutor(pool_size) as pool:
+                futures = []
+                for p in range(1, partitions + 1):
+                    futures.append(
+                        pool.submit(
+                            _run,
+                            self.executable_path,
+                            self.force_flag,
+                            self.scale_flag,
+                            str(scale_factor),
+                            self.partitions_flag,
+                            str(partitions),
+                            self.current_segment_flag,
+                            str(p),
+                            cwd=self.executable_path.parent,
+                        )
+                    )
+                for f in tqdm.tqdm(
+                    concurrent.futures.as_completed(futures), total=partitions
+                ):
+                    f.result()
+
+            # Move the new files to out_dir
+            out_dir = pathlib.Path(out_dir).resolve()
+            for table_name in self.table_names:
+                table_output_dir = out_dir / f"{table_name}.csv"
+                table_output_dir.mkdir()
+                for p in range(1, partitions + 1):
+                    old_file = (
+                        self.executable_path.parent
+                        / self._get_partitioned_filename(table_name, p, partitions)
+                    )
+                    new_file = table_output_dir / f"part-{p}.csv"
+                    # Not all tables will have enough rows for the full nr of partitions
+                    if old_file.exists():
+                        shutil.move(old_file, new_file)
+                        # reset permissions to read-only
+                        os.chmod(new_file, 0o444)
+                        log.debug(f"Created {new_file}")
+                os.chmod(table_output_dir, 0o555)
+                log.debug(f"Created all partitions for {table_name}")
 
     def _make_executable(self):
         """Clone the repo and build the executable."""
@@ -184,7 +240,7 @@ class DBGen(_TPCBuilder):
     scale_flag = "-s"
     partitions_flag = "-C"
     current_segment_flag = "-S"
-    file_extension = ".tbl"
+    file_extension = "tbl"
     table_names = tpc_table_names["tpc-h"]
     repo_uri = "https://github.com/electrum/tpch-dbgen.git"
     repo_commit = "32f1c1b92d1664dba542e927d23d86ffa57aa253"
@@ -197,6 +253,13 @@ class DBGen(_TPCBuilder):
             return self.repo_build_path / "Debug" / "dbgen.exe"
         else:
             return self.repo_build_path / "dbgen"
+
+    def _get_partitioned_filename(self, table_name, partition, total_partitions):
+        if table_name == "nation":  # TODO: and total_partitions > some_number
+            return f"nation.{self.file_extension}"
+        if table_name == "region":
+            return f"region.{self.file_extension}"
+        return f"{table_name}.{self.file_extension}.{partition}"
 
     def _build_executable_unix(self):
         """Build the executable using 'make' on a UNIX-based system."""
@@ -235,7 +298,7 @@ class DSDGen(_TPCBuilder):
     scale_flag = "-SCALE"
     partitions_flag = "-PARALLEL"
     current_segment_flag = "-CHILD"
-    file_extension = ".dat"
+    file_extension = "dat"
     table_names = tpc_table_names["tpc-ds"]
     repo_uri = "https://github.com/gregrahn/tpcds-kit.git"
     repo_commit = "5a3a81796992b725c2a8b216767e142609966752"
@@ -248,6 +311,9 @@ class DSDGen(_TPCBuilder):
             return self.repo_build_path / "dsdgen.exe"
         else:
             return self.repo_build_path / "dsdgen"
+
+    def _get_partitioned_filename(self, table_name, partition, total_partitions):
+        return f"{table_name}_{partition}_{total_partitions}.{self.file_extension}"
 
     def _build_executable_unix(self):
         """Build the executable using 'make' on a UNIX-based system."""
